+++ CHANGELOG.md
@@ -12,6 +12,14 @@ CHANGELOG
   # Send actions to the server
   curl -XPOST localhost:6266 -d 'reload(seq 100)+change-prompt(hundred> )'
   ```
+- Added scrollbar on the main search window
+  ```sh
+  # Hide scrollbar
+  fzf --no-scrollbar
+
+  # Customize scrollbar
+  fzf --scrollbar ┆ --color scrollbar:blue
+  ```
 - New event
     - Added `load` event that is triggered when the input stream is complete
       and the initial processing of the list is complete.